arm64: dts: rockchip: specify on-board GPIO LEDs for axon
Axon contains 2 on-board LEDs which are being handled from the
PCA9554 GPIO expander.

One GPIO is configured as power led and another is used as
status LED for the given function from the userspace.

By default second LED will be in heartbeat trigger mode.

Both of the LEDs will be always on.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>