arm64: dts: rockchip: add two usb2 support for axon
Enable USB2 support for axon in EHCI and OHCI both modes.
Both USB-A ports on the axon are USB2 ports.

Make regulator for the USB2 from the 5V rail of the board.
Regulator can be controlled by the given GPIO. For that
pinctrl node has been added.

Both USB3 phy goes to Type-C, it will be added in follow up
patches.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>